requirements.txt
aniso8601==1.2.0
bcrypt==3.1.1
blinker==1.4
cffi==1.9.1
click==6.6
colorama==0.3.7
coverage==4.2
decorator==4.0.10
Flask==0.11.1
Flask-HTTPAuth==3.2.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
Flask-User==0.6.8
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.9
passlib==1.6.5
pbr==1.10.0
pycparser==2.17
pycrypto==2.6.1
python-dateutil==2.5.3
pytz==2016.7
six==1.10.0
SQLAlchemy==1.1.3
sqlalchemy-migrate==0.10.0
sqlparse==0.2.2
Tempita==0.5.2
Werkzeug==0.11.11
WTForms==2.1
